+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Remove some main menu entries (events, CADA, Etalab)
+- Use a new SVG logo
+- Apply changes from [uData 1.0.0](https://pypi.python.org/pypi/udata/1.0.0#changelog)
 
 ## 0.9.1 (2017-01-10)